requirements.txt
PyYAML==6.0
captum==0.3.1
click==7.1.2
dash-bootstrap-components==0.12.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash==1.19.0
dill==0.3.3
gensim==4.1.0
imbalanced-learn==0.8.0
jinja2==2.11.3
matplotlib==3.5.1
morphio==3.3.3
morphology-workflows==0.2.0
networkx==2.5
neurom==3.1.0
numpy==1.20.3
pandas==1.3.4
plotly==5.4.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.0
shap[plots]==0.39.0
tmd==2.1.0
torch==1.7.1
tqdm==4.53.0
typing-extensions==4.1.1
umap-learn==0.5.1
xgboost==1.4.2
